+++ crates/xtask/src/cmd/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod audit;
 pub(crate) mod compare;
 pub(crate) mod debug;
 pub(crate) mod generate;
@@ -6,6 +7,7 @@ pub(crate) mod overrides;
 pub(crate) mod snapshots;
 pub(crate) mod verify;
 
+pub(crate) use audit::*;
 pub(crate) use compare::*;
 pub(crate) use debug::*;
 pub(crate) use generate::*;